moved screenshot capturing code
