+++ lib/utils.js
@@ -32,6 +32,7 @@
 
 const fs = require('fs');
 const path = require('path');
+const sprintf = require('sprintf-js').sprintf;
 
 const promiseLoop = function (condition, body) {
   var loop = function () {
@@ -67,6 +68,27 @@ const getSelectorDimensions = function (page, selector) {
   }, selector);
 };
 
+const makeFileNameConverter = function (config) {
+  var fileNameConverter = config.fileNameConverter;
+  if (!fileNameConverter) {
+    if (config.outputPattern) {
+      fileNameConverter = function (num) {
+        return sprintf(config.outputPattern, num);
+      };
+    } else if (config.frameProcessor && !config.outputDirectory) {
+      fileNameConverter = function () {
+        return undefined;
+      };
+    } else {
+      fileNameConverter = function (num, maxNum) {
+        var outputPattern = '%0' + maxNum.toString().length + 'd.png';
+        return sprintf(outputPattern, num);
+      };
+    }
+  }
+  return fileNameConverter;
+};
+
 const makeFileDirectoryIfNeeded = function (filepath) {
   var dir = path.parse(filepath).dir, ind, currDir;
   var directories = dir.split(path.sep);
@@ -82,5 +104,6 @@ module.exports = {
   promiseLoop,
   getBrowserFrames,
   getSelectorDimensions,
+  makeFileNameConverter,
   makeFileDirectoryIfNeeded
 };
